Buxfix for the a-b-c case. Third derived reference from an original only cleand up the pointers to "currently active record" for the previous, instead of all previous, taxa.
New feature to support a-b-a' case, where we move a taxon from one genus to another and then
back. This would create a new record for a', instead of updating the old record for a. There
is now a modal dialog box that lets the user choose the next step.