(regression) Removed 'edit references in place' feature.
First step in cleaning up the ReferencesController and related code. Too hard to maintain. Adding and editing references in-place is a really neat feature, but this code is to 99% a duplication (with slightly different attribute/variable names) of the original add/edit references code, which has its own set of problems. Let's try to clean up the code and then make a decision if we want to restore this feature. JavaScripts/stylesheets have not been touched.