(regression) Removed special CSS for references with downloadable documents.
I always wondered why some references were grayed-out, perhaps they were not up to date or invalid? This is not what I expected, and the information is already contained in the fact that linked references have PDF links. No objections for restoring if this feature is required, but just making the PDF link more visible, eg linking the whole title (which we probably should do anyways) would be better affordance.