[Russian roulette] Removed IE-only stylesheet.
I haven't actually tested this. But 1) A new 'img.help_icon' for the tooltips is defined elsewhere and I doubt "top: -100px" makes sense even in IE. 2) That leaves only a padding fix x 2, which likely is fixed in IE > 6. 3) We need to use an HTML normalizer if we are to do this at all.