Split template
Slightly different concerns: the 'taxon operations' (ugg naming is hard -- I tried with left/right, but that was even worse) or "add taxon or add tribe or convert to subspecies, or elevate to species" section is for persisted taxa only. The 'save section' is for all taxa. To be improved further.

The local variable 'form' was not used; removed. Also discovered that @possible_homonym is not declared anywhere.