Removed redundant div
It looks like an hr element in the view, because it's not nested correctly. The 'unless' statement needs to be indented an additional level, but after trying that I decided that it looks better without a large panel stretching the whole page considering how small the nested Author Names panel is (and it already has its own styling/borders). This also makes it clear that 'button_container' from the Coffee is not used.