Removed the last of jQuery UI's button() functions
Change: Like commit message title + updated the Coffee to not select on spans which were generated by that function.

Reason: Less coupled to jQuery UI (and its theme). Easier to maintain, and this commit makes it obvious how hard it is to work with code that depends on JavaSscript-generated CSS/elements (the selector that was modified in this commit also selected on hard-coded CSS classes, but we're not ready to remove these yet). Also makes styling easier.