Fixed the old autoloading issue.
The one involving Parsers::AuthorParser. The Citrus grammar files are probably not reloaded during dev (which was the whole reason for the now removed code), but we're not working on that and since it crashes the site in dev all the time I've not even bothered to check if they are autoloaded (or if they even were autoloaded before this commit).