Fix autoloading?
Didn't work in c8dfb59b873d68cc0e252e87d2f23ec8354d8890. It seemed like it did but now I cannot reproduce that at all.

Seems to work better using require_depedency (Ruby's require does not play wery well with Rail's autoloading).

There are still a couple of require in the code, but let's see how this goes and if we hit the same autoloading issue again in other parts of the code and using require_dependency solves that, then at least we can be pretty sure *why* we had these issues (or at least one of the reasons for autoloading not working).