Nil check to avoid crash
To reproduce: create a genus, create a species in that genus. Now there's two changes on the changes page (change_type == create).

So far all good.

Then delete the genus, which also deleted the species. Now we have three changes on the change page: the two create events and a delete event (the genus).

The species is not listed as a deleted taxon, because it's included in its parent's delete event.

However, the species is still in the list (as the initial create event). The data we created the species with is visible in both the delete event and create events, because Change#get_most_recent_valid_taxon reifies the deleted taxon. But it does not reify the parent genus.

This is where it crashed when trying to display the species