Tweaked rank_to_create (TaxaController)
This is cleaner, IMO, because it makes it clear that rank_to_create is only used when creating new taxa.

Some of my notes:
new/create
The new action always has access to params[:rank_to_create] (it's not possible to create new taxa without following a link that includes it in the url). The create action also always has access to it, becuase it's set in the form in the new action. That's why we can just move the fetching of the param to #get_taxon_for_create (which is the only method that uses this param / instance variable).

edit/update do not use this variable.

It was previously used for the "Add subfamily/genus/species" button, but that was removed in 2d1ee7aefcf5f7da31e120fbe5b6ac1e339d3c13.