Remove id instance variable + some notes on what I'm trying to accomplish
I did not study TaxonMother until very recently, and it has a couple of issues. It's possible that it used to make more sense, but it's hard to tell now after so much dead code has been removed. Right now this class is responsible for too many different concerns, is non-intuitive in places, contains unnecessary abstractions and defines code too far away from where it's used. Some initial thoughts that may or may not be true:

* `save_taxon` looks like a class method pretending to be an method?

* The factory method `create_taxon` requires first instantiating the class?

* `load_taxon` also requires an instance of the class (and this is the only method that makes use of @id)?

* That kind of makes `initialize` unnecessary, because all other methods are more helpers/utility/class methods defined as methods?

* Many methods belong more in other classes, eg `build_children` in TaxaController, `delete_taxon` in Taxon (which admittedly already is crowded -- we have to work on that later after deciding what to do with TaxonMother), etc, etc.

Not sure, to be investigated!