[wip] modified SaveTaxon to work on the instance variable
Set @taxon in the constructor and use that. Don't allow initializing the class with an empty or random id (or string) like we previously did because some code wanted to make calls to "method class methods".

This is what I meant with that it looks like a class method -- it did not really care about any previous state and just initialized @taxon to whatever was sent into the metod.

Eg `@taxon.id` in `mother = TaxonMother.new @taxon.id` wasn't used, and the same mother was later reused for a different taxon.