Added indexes to versions
I'm not really good with databases/indexes, but this improves performance a lot.

Clear cache and browse the a couple of change pages before appying the migration. Migrate, reset cache again and compare. Page loads are still very slow, but some of that has to do with other things.

Busting the cache:
$rails dbconsole
mysql>RESET QUERY CACHE;

The taxt_items are from an expermimental branch and snuck into the schema in 6e38e57afd636827c8f9c5142b86a9f0893f2240.