Prepare next series of commits
Admittedly rather clunkly. To be improved.

The old problem: The taxon browser takes up too much space on the screen. Moving it to the bottom would effectively hide it on long pages (and the root page happens to be very long).

Solution: Hide it by default (there's already code to toggle show/hide).

New problems: 1) All tests browsing the catalog will become dependent on JS --> suite becomes painfully slow (because we cannot use database transaction for rolling back the db) --> developers becomes unhappy with their lives in general. 2) We have to rewrite all test.

To counter "1" I'm ok with this test-specific code. To counter "2" I'm ok with this solution until we're ready to rewrite to tests.

More stuff: `AntCat.environment` was not used. JS is probably better than Coffee in application.haml for performance reasons. The [future] code that hides the browser will be enbled by default in prod and dev, but disabled in test unless the scenario is tagged with @enable_taxon_browser_autohide.