Added keep [taxon browser] open option
And default to false.

What we do not want: users becoming angry because they have to *open* the taxon browser all the time because they want to click loads of links from the browser.

What we also do not want: users having to *close* the taxon browser all the time because it is annoying to have around when not used.

What we want: this!! which is: keep open = continue to have to taxon browser open unless the user has closed it.

I'm pretty sure this can be made A LOT easier. TODO figure that out once related code is more complete.