(big change) Re-implemented the taxon browser
Problem with the *code*: it is very hard to work with.

Try to do any of these and "very hard" is more like "impossible" + 100 conditionals:
* Toggle display valid only.
  * And let the user know when to toggle to see the hidden taxa instead of showing nothing.
* Always show tribes.
* List all subspecies of a species in the index.
  * But not if there are none.
* Show all Formicidae genera.
* Only show the incertae sedis (previously "none") links if there are taxa of this type.
* Hide empty panels.
* Smart(er) conditionals for collapsing panels (ie "close that panel if such and such").
* Implement filtering besides "valid only" etc [the only mentioned item not done is this commit.]

Problem with the UX. We want to be able to:
* As above.
* Important: not force visitors to reload the page to toggle show/hide tribe/subgenera, in particular when there are no tribes/subgenera so that the user has to toggle back to see the species again.
* Also important: show valid only.

There are some oddities left in the code and some some new oddities, but this is to be improved somehow, hopefully. This commit removed the subgenus panel (this rank is not used that often anyways) so maybe we should make it easier to browse them from the taxon descriptions. This change also makes is possible to add new panels for eg species groups and subtribes (not that that is likely to happen any time soon).

Sorry for this huge commit; it simply was not possible to change this code in smaller steps / would had been confusing anyways / would take 9623423823798 years to run the tests.

Not done:
* Properly marking incertae sedis links as selected when child taxa are selected.
* Ecitoninae et al. are not visible by default (ie the subfamilies that were synonymized under Dorylinae in 2014) -- but now that "show valid only" is a thing it's possible to see them that way, so we do not really need it, and we're also not lying which we would if they were visible when "show valid only" was active.

  Bonus code: Use this to restore the old behavior (recently synonymized)
  if session[:show_valid_only]
    children =  if taxon.is_a? Family
                  children.where.not(status: "unavailable")
                else
                  children.valid unless taxon.is_a? Family
                end
  end