Removed Rank + comments
Note about DRY-ness:

Didn't I just take a central class, where we had a single place for defining this behavior and then just distributed it in multiple places, making it impossible to keep it up-to-date?

----------

`rank_options_for_select` needs all the rank except family, plus an "All" option, and it doesn't care about "uncommon" ranks. Format: Singular capitalized => Plural capitalized. Simply listing them in an hash makes this so much easier to read, and it avoids re-implementing "options_for_select" as "option_for_select" (but with the arguments in reverse order). DRY = OK.

---------

`add_taxon_button` wants to know the next child rank, but the "uncommon" ranks are broken because we want to be able to add more than one ranks to some ranks (eg tribe and genus in subfamilies), so we need special logic for that (the 'add tribe' code is found in `add_tribe_button` and we have no button for 'add subgenus'). DRY = OK because this does not fit really well in a simple hierarchy.

--------

Code related to setting the parent is just too complicated already, and this change makes it possible to sort that out. Currently we have 1.) `#parent=`s in the Taxon classes, 2.) `#update_parent`s in the Taxon classes, 3.) `#update_parent` in TaxaController, 4.) setting the fields directly in the taxon form / other code, and 5.) some before_filters setting some fields in some classes.

This was the only code that used the "next parent rank" (except for a call in a HOL class that I think is going to be removed?). DRY = this code was never DRY in the first place.

----