[ReferenceForm+ReferencePanel] Remove unused/deprecated/duplicated JS
After removing the two calls to `reference_panel()` there are no live references to `ReferenceForm` or `ReferencePanel`.

I'm not 100% sure that this was ever used; perhaps is was, or perhaps it wasn't but it was supposed to replace code now in "reference_edit_autocompletion.js" and "references.coffee". See also the note in the diff regarding "This seems the be the original JavaScript that ...".

We should remove this because:
1) it's not used
2) all of the code is duplicated elsewhere, in live code
3) it makes it harder to understand the already hard-to-understand CoffeeScript subdir / everything
4) it makes it harder to change unrelated code that happens to share the same CSS selectors (this is how I ended up here)

-------------

Some notes from the investigation:

Main suspect: `ReferenceForm` alias "widgets/reference_form.coffee".

Contains code for reference autocompletion and reference tabs -- none of which are used in the reference picker.

The file defines `AntCat.ReferenceForm`, which is only referenced once, in "widgets/reference_panel.coffee".

The only reference to `ReferencePanel` is code inside the same file, in the jQuery plugin `$.fn.reference_panel`

There are only two calls to `reference_panel()`, both of which we do not need/want.