(feed) link related taxa in + comments
We may want to get smarter about this. Adding a polymorphic recipient to the activity would make it possible to tie eg edited reference sections to the taxon they belong to (which makes querying possible); but then we also have the synonym relationships which would require two recipients. We probably need a new first-class model with a join table anyways; that way we could link tons of items which could be queried against from the db, and show that "mini-feed" on individual pages (really a change log of events related to that taxon).

Example of why we cannot use a recipient (or two) all the way:
* Change the parent of a genus (with 20 species) --> 20 recipients?