(feedback) Added model/controller/modal
GH issue: #123

This change replaces the feedback email link with a modal that allows registered and unregistered users to sent feedback/corrections. The old email link is still in the GUI (visible after opening the feedback modal).

Feature is already completed, just commiting it in smaller steps.

In the next commits:
* ActiveAdmin
* Feedback sent notice
* Actually sending the email
* Tests