Moved the footer (ie partner links) to the about page.
It's too hard to make a resizable taxon browser work with this in the catalog. Keeping it on all pages except the catalog could also work, but I never really liked the idea of always showing these anyways, so they have been banished to the about page at least for now.