Cucumber/Changes: don't feature test taxon representation
By which I mean that "I should see the locality Africa in the changes" does not mean "confirm that the change stores previous/current values for the locality field'.

It just test a representation of the taxon as it currently stands, so it's pretty much testing "When I go to the catalog page I should see Africa".